测试v1.0

+++ src/main/java/com/sziit/noteassistant/controller/AudioController.java
@@ -10,9 +10,13 @@ import com.sziit.noteassistant.pojo.entity.Note;
 import com.sziit.noteassistant.pojo.entity.User;
 import com.sziit.noteassistant.service.AudioService;
 import com.sziit.noteassistant.service.NoteDetailService;
+import com.sziit.noteassistant.utils.JwtUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -29,39 +33,69 @@ import java.time.LocalDateTime;
 @RestController
 public class AudioController {
 
+    protected Log logger = LogFactory.getLog(this.getClass());
     @Autowired
     private AudioService audioService;
 
     @GetMapping("findAudio")
     @ApiOperation(value = "查询音频")
+    @Transactional(rollbackFor = {Exception.class,RuntimeException.class})
     public Object findAudio(@RequestParam Integer aId) {
+        User user = JwtUtils.getUserBytoken();
+        if(user.getUsername()==null){
+            logger.error("未登录");
+            return new ResultVo(ResultCode.UNAUTHORIZED);
+        }
         return new ResultVo(audioService.selectAudioByAid(aId));
     }
 
     @PostMapping("addAudio")
     @ApiOperation(value = "添加音频")
+    @Transactional(rollbackFor = {Exception.class,RuntimeException.class})
     public Object addAudio(@RequestBody Audio audio) {
+        User user = JwtUtils.getUserBytoken();
+        if(user.getUsername()==null){
+            logger.error("未登录");
+            return new ResultVo(ResultCode.UNAUTHORIZED);
+        }
         audioService.addAudio(audio);
         return new ResultVo(audioService.selectAudioOne(audio));
     }
 
     @DeleteMapping("delAudio")
     @ApiOperation(value = "删除音频")
+    @Transactional(rollbackFor = {Exception.class,RuntimeException.class})
     public Object deleteAudio(@RequestParam Integer aId) {
+        User user = JwtUtils.getUserBytoken();
+        if(user.getUsername()==null){
+            logger.error("未登录");
+            return new ResultVo(ResultCode.UNAUTHORIZED);
+        }
         audioService.deleteAudio(aId);
         return new ResultVo(ResultCode.SUCCESS);
     }
 
     @PutMapping("updateAudio")
     @ApiOperation(value = "修改音频")
+    @Transactional(rollbackFor = {Exception.class,RuntimeException.class})
     public Object updateAudio(@RequestBody Audio audio){
+        User user = JwtUtils.getUserBytoken();
+        if(user.getUsername()==null){
+            logger.error("未登录");
+            return new ResultVo(ResultCode.UNAUTHORIZED);
+        }
         return new ResultVo(audioService.updateAudio(audio));
     }
 
     @DeleteMapping("delAudios")
     @ApiOperation(value = "删除多个音频")
+    @Transactional(rollbackFor = {Exception.class,RuntimeException.class})
     public Object deleteAudios(@RequestParam Integer[] aIds) {
-        JSONObject jsonObject = new JSONObject();
+        User user = JwtUtils.getUserBytoken();
+        if(user.getUsername()==null){
+            logger.error("未登录");
+            return new ResultVo(ResultCode.UNAUTHORIZED);
+        }
         boolean result = audioService.deleteAudios(aIds);
         if (result) {
             return new ResultVo(ResultCode.SUCCESS);
